chore: api 함수들 위치 이동(코로케이션 적용)
