+++ src/app/(main)/api/index.ts
@@ -1,0 +1,1 @@
+export { default as getSentences } from './getSentences';
