+++ src/app/book/[id]/api/getBook.ts
@@ -1,0 +1,8 @@
+import { fetchAPI } from '@/api/fetcher';
+import type { Book } from '@/types';
+
+const getBook = async (bookId?: string) => {
+  return fetchAPI<Book>(`/books/${bookId}`);
+};
+
+export default getBook;
