+++ src/app/book/[id]/api/index.ts
@@ -1,0 +1,1 @@
+export { default as getBook } from './getBook';
