+++ src/app/edit/sentence/components/SentenceEditContainer/api/index.ts
@@ -1,0 +1,2 @@
+export { default as createSentence } from './createSentence';
+export { default as updateSentence } from './updateSentence';
