+++ src/app/edit/sentence/components/SentenceEditContainer/api/type.ts
@@ -1,0 +1,7 @@
+import type { Book } from '@/types';
+
+export type CreateSentenceParams = {
+  book: Book;
+  content: string;
+};
+export type UpdateSentenceParams = CreateSentenceParams & { id: string };
