+++ src/app/my/like/components/MyLikeList/api/getUserLikes.ts
@@ -1,0 +1,13 @@
+import { fetchAPI } from '@/api/fetcher';
+import type { PaginationResult, Sentence } from '@/types';
+
+const SENTENCE_PAGE_LIMIT = 24;
+
+const getUserLikes = async (page: string) => {
+  return fetchAPI<PaginationResult<Sentence>>(
+    `/users/me/likes?page=${page}&limit=${SENTENCE_PAGE_LIMIT}`,
+    { cache: 'no-store' },
+  );
+};
+
+export default getUserLikes;
