+++ src/app/my/like/components/MyLikeList/api/index.ts
@@ -1,0 +1,1 @@
+export { default as getUserLikes } from './getUserLikes';
