+++ src/app/my/sentence/components/MySentenceCardButtons/api/deleteSentence.ts
@@ -1,0 +1,7 @@
+import { fetchAPI } from '@/api/fetcher';
+
+const deleteSentence = (sentenceId: string) => {
+  return fetchAPI(`/sentences/${sentenceId}`, { method: 'DELETE' });
+};
+
+export default deleteSentence;
