+++ src/app/my/sentence/components/MySentenceCardButtons/api/index.ts
@@ -1,0 +1,1 @@
+export { default as deleteSentence } from './deleteSentence';
