+++ src/app/my/sentence/components/MySentenceList/api/index.ts
@@ -1,0 +1,1 @@
+export { default as getUserSentences } from './getUserSentences';
