+++ src/app/my/setting/components/UserActionButtons/api/deleteUser.ts
@@ -1,0 +1,10 @@
+import { fetchAPI } from '@/api/fetcher';
+import { auth } from '@/lib/firebase';
+import { signOut } from '@firebase/auth';
+
+const deleteUser = async () => {
+  await fetchAPI('/users/withdrawal', { method: 'DELETE' });
+  await signOut(auth);
+};
+
+export default deleteUser;
