+++ src/app/my/setting/components/UserActionButtons/api/index.ts
@@ -1,0 +1,1 @@
+export { default as deleteUser } from './deleteUser';
