+++ src/components/molecules/LikeButton/api/index.ts
@@ -1,0 +1,1 @@
+export { default as toggleSentenceLike } from './toggleSentenceLike';
