+++ src/components/molecules/LoginButton/api/index.ts
@@ -1,0 +1,1 @@
+export { default as loginWithGoogle } from './loginWithGoogle';
